src/python/FCMPrinterStack/requirements.txt
fastapi
uvicorn[standard]
pycups